Rename files to follow underscore naming convention
