+++ lib/deck_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firestore_again/cubit/bigStates.dart';
-import 'package:firestore_again/cubit/logicCubit.dart';
+import 'package:firestore_again/cubit/big_states.dart';
+import 'package:firestore_again/cubit/logic_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
